// src/app/main/shared/models/Complicaciones.ts
export interface Complicaciones{
  swInfartoAgudoMiocardio: boolean;
  infartoAgudoMiocardio: string | null;
  swAccidenteCerebroVascular: boolean;
  accidenteCerebroVascular: string | null;
  swNefropatiaDiabetica: boolean;
  nefropatiaDiabetica: string | null;
  swRetinopatiaHipertensiva: boolean;
  retinopatiaHipertensiva: string | null;
  swEnfermedadRenalCronica: boolean;
  enfermedadRenalCronica: string | null;
  swNeuropatiaPeriferica: boolean;
  neuropatiaPeriferica: string | null;
  swInsuficienciaCardiaca: boolean;
  insuficienciaCardiaca: string | null;
  swAmputacion: boolean;
  amputacion: string | null;
  swHigadoGraso: boolean;
  higadoGraso: string | null;
  swCirrosis: boolean;
  cirrosis: string | null;
  swEpocEpid: boolean;
  epocEpid: string | null;
}